+++ 404.php
@@ -4,7 +4,7 @@
  *
  */
 ?>
-<link href="<?php echo get_template_directory_uri() ?>/css/fonts.css"
+<link href="<?php echo get_template_directory_uri() ?>/assets/dist/css/fonts.css"
       rel="stylesheet">
 <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap"
       rel="stylesheet">